GPL 1.1.2.3_224

+++ apps/public/asuswebstorage/api.h
@@ -9,7 +9,6 @@
 #define DOWNLOAD 2
 #define UP_EXCEP_FAIL 3
 
-//#define DEBUG 1
 #define SYSTEM_LOG 0
 #define my_free(x)  if(x) {free(x);x=NULL;}
 #define TREE_NODE_ENABLE 0
@@ -100,17 +99,13 @@ typedef void (*processing)(double current, double total);
 
 extern char mount_path[NAMESIZE];
 extern char cloud_path[NAMESIZE];
-//char temp_path[NAMESIZE];
 extern char asus_path[NAMESIZE];
 extern char log_path[NAMESIZE];
 extern char xml_path[NAMESIZE];
-//char mount_path[NAMESIZE];
 extern char sync_path[NAMESIZE];
 extern char system_log[NAMESIZE];
 extern char general_log[NAMESIZE];
-//char tree_log[NAMESIZE];
 extern char confilicted_log[NAMESIZE];
-//char temp_file[NAMESIZE];
 extern char up_item_file[NAMESIZE];
 extern char down_item_file[NAMESIZE];
 
@@ -118,9 +113,7 @@ extern char down_item_file[NAMESIZE];
 char down_item_temp_file[NAMESIZE];
 #endif
 
-//char all_local_item_file[NAMESIZE];
 extern char up_excep_fail_file[NAMESIZE];
-//char up_limit_file[NAMESIZE];
 extern char gateway_xml[NAMESIZE];
 extern char get_user_state_xml[NAMESIZE];
 extern char token_xml[NAMESIZE];
@@ -135,21 +128,10 @@ extern char move_xml[NAMESIZE];
 extern char remove_xml[NAMESIZE];
 extern char update_xml[NAMESIZE];
 extern char get_entry_info_xml[NAMESIZE];
-//char set_mark_xml[NAMESIZE];
-//char get_change_files_xml[NAMESIZE];
-//char get_uploads_xml[NAMESIZE];
-//char get_share_code_xml[NAMESIZE];
-//char del_share_code_xml[NAMESIZE];
-//char get_share_entry_xml[NAMESIZE];
-//char check_pwd_xml[NAMESIZE];
-//char cmp_pwd_xml[NAMESIZE];
 extern char get_change_seq_xml[NAMESIZE];
 extern char init_upload_xml[NAMESIZE];
 extern char resume_upload_xml[NAMESIZE];
 extern char finish_upload_xml[NAMESIZE];
-//char get_resize_photo_xml[NAMESIZE];
-//char get_full_txt_xml[NAMESIZE];
-//char get_video_snapshot_xml[NAMESIZE];
 extern char trans_excep_file[128];
 extern char system_token[256];
 
@@ -167,81 +149,68 @@ int obtainToken(char *user,char *pwd,struct asus_config *cfg,int first);
 
 Getinfo *getInfo(char *username,char *server);
 Getmysyncfolder *getMySyncFolder(char *username);
-int obtainSyncRootID(char *user);
+long long int obtainSyncRootID(char *user);
 Getpersonalsystemfolder *getPersonalSystemFolder(char *username,char *filename);
-int GetMyRecycleID(char *username,char *filename);
+long long int GetMyRecycleID(char *username,char *filename);
 Getuserstate *getUserState(char *user,char *server);
 int CheckUserState(char *user,char *server);
 int get_max_upload_filesize(char *username);
 
-Browse *browseFolder(char *username,int id,int issibiling,int pageno);
+Browse *browseFolder(char *username,long long int id,int issibiling,int pageno);
 int my_parse_browse_xml(const char *filename,Browse *br);
-//Browse *GetServerList(char *username,int id,int issibiling);
-//int browseFolder(char *username,int id,int issibiling,Browse *br);
-Propfind *checkEntryExisted(char *username,int parentID,char *filename,char *type);
-Createfolder *createFolder(char *username,int parentID,int isencrpted,char *name);
+Propfind *checkEntryExisted(char *username,long long int parentID,char *filename,char *type);
+Createfolder *createFolder(char *username,long long int parentID,int isencrpted,char *name);
 
-Operateentry *renameEntry(char *username,int id,int isencrpted,char *newname,int isfolder);
-Moveentry *moveEntry(char *username,int id,char *name,int parentID,int isfolder,int pre_pid);
-Operateentry *removeEntry(char *username,int id,int ischildonly,int isfolder,int pid);
-int updateEntryAttribute(char *username,int id,int parentID,int isencrpted,int isfolder);
+Operateentry *renameEntry(char *username,long long int id,int isencrpted,char *newname,int isfolder);
+Moveentry *moveEntry(char *username,long long int id,char *name,long long int parentID,int isfolder,long long int pre_pid);
+Operateentry *removeEntry(char *username,long long int id,int ischildonly,int isfolder,long long int pid);
+int updateEntryAttribute(char *username,long long int id,long long int parentID,int isencrpted,int isfolder);
 
-Getentryinfo *getEntryInfo(int isfolder,int entryid);
-//int setEntryMark(int isfolder,int entryid,int markid);
-//int getLatestChangeFiles(char *username,int top,int targetroot,int sortdirection);
-//int getLatestUploads(char *username,int top,int targetroot,int sortdirection);
+Getentryinfo *getEntryInfo(int isfolder,long long int entryid);
+
 
 /*shared files API*/
-//int getShareCode(char *username,int entryType,int entryID,char *password,int actionType);
-//int deleteShareCode(char *username,int entryType,int entryID,char *password);
-//int getSharedEntries(char *username,int kind,int pagesize,int sortby,int sortdirection,char *firstentrybound);
-//int checkPassword(char *username,char *suri);
-//int comparePassword(char *username,int isfolder,int ffid,char *password);
-Changeseq *getChangeSeq(int folderid);
+Changeseq *getChangeSeq(long long int folderid);
 
 /* upload file*/
 int sha512(char *filename,char *checksum);
-Initbinaryupload  *initBinaryUpload(char *filename,int parentID,char *transid,int fileID);
+Initbinaryupload  *initBinaryUpload(char *filename,long long int parentID,char *transid,long long int fileID);
 Resumebinaryupload *resumeBinaryUpload(char *filename, Initbinaryupload *ibu);
 Finishbinaryupload *finishBinaryUpload(Initbinaryupload *ibu);
-int uploadFile(char *filename,int parentID,char *transid,int fileID);
-//int check_exist_on_server(char *username,char *filename,int parentID);
+int uploadFile(char *filename,long long int parentID,char *transid,long long int fileID);
 
 /*download file*/
 int update_local_file_attr(Fileattribute *attr,char *filename);
-int downloadFile(int fileID,char *filename,long long int size,int ismodify,Fileattribute *attr);
+int downloadFile(long long int fileID,char *filename,long long int size,int ismodify,Fileattribute *attr);
 
-int getResizedPhoto(int photoFileID,int sizeType,int preview);
-int getFullTextCompanion(int fileID,int preview,int key);
-int getVideoSnapshot(int filmID,int preview);
-
-//int syncFolder(char *parentfolder,Browse *browse,Local *local);
+int getResizedPhoto(long long int photoFileID,int sizeType,int preview);
+int getFullTextCompanion(long long int fileID,int preview,int key);
+int getVideoSnapshot(long long int filmID,int preview);
 
 int handle_error(int code,char *type);
-int getParentID(char *path);
+long long int getParentID(char *path);
 long int check_server_space(char *username);
-//int write_log(Browse *bs);
 int write_log(int status,char *message,char *filename);
 int write_finish_log();
 int write_system_log(char *action,char *name);
 int write_confilicted_log(char *prename,char *confilicted_name);
 int write_trans_excep_log(char *fullname,int type,char *msg);
-int sync_all_item(char *dir,int parentID);
-int sync_all_item_uploadonly(char *dir,int parentID);
+int sync_all_item(char *dir,long long int parentID);
+int sync_all_item_uploadonly(char *dir,long long int parentID);
 int add_all_download_only_socket_list(char *cmd,const char *dir);
 int add_all_download_only_dragfolder_socket_list(const char *dir);
-int handle_rename(int parentID,char *fullname,int type,char *prepath,int is_case_conflict,char *pre_name);
-int handle_createfolder_fail_code(int status,int parent_ID,char *path,char* fullname);
-int handle_upload_fail_code(int status,int parent_ID,char* fullname,const char *path);
+int handle_rename(long long int parentID,char *fullname,int type,char *prepath,int is_case_conflict,char *pre_name);
+int handle_createfolder_fail_code(int status,long long int parent_ID,char *path,char* fullname);
+int handle_upload_fail_code(int status,long long int parent_ID,char* fullname,const char *path);
 int handle_delete_fail_code(int status);
-int handle_rename_fail_code(int status,int parentID,char *fullname,char *path,int isfolder);
-int handle_move_fail_code(int status,char *path,char *fullname,int parentID,char *prepath,int entryID,int isfolder);
-int create_server_folder_r(const char *path);
-int upload_entry(char *fullname,int parent_ID,char *path);
+int handle_rename_fail_code(int status,long long int parentID,char *fullname,char *path,int isfolder);
+int handle_move_fail_code(int status,char *path,char *fullname,long long int parentID,char *prepath,long long int entryID,int isfolder);
+long long int create_server_folder_r(const char *path);
+int upload_entry(char *fullname,long long int parent_ID,char *path);
 
 int StrToHex(char *src,int len);
 
-int IsEntryDeletedFromServer(int fileID,int isfolder);
+int IsEntryDeletedFromServer(long long int fileID,int isfolder);
 int obtain_token_from_file(const char *filename,Aaa *aaa);
 
 int count_call_api(const char *url);
